feat: мидлвэр с id удален, убрана возможность удалять карточки других пользователей

+++ controllers/cards.js
@@ -1,7 +1,8 @@
 const Card = require("../models/card");
 const Codes = require("../utils/codes");
+const { BadRequestErr, ForbiddenErr, NotFoundErr } = require("../utils/errors");
 
-const createCard = async (req, res) => {
+const createCard = async (req, res, next) => {
   try {
     const owner = req.user._id;
     const { name, link } = req.body;
@@ -9,43 +10,34 @@ const createCard = async (req, res) => {
     return res.status(Codes.Created).json(card);
   } catch (err) {
     if (err.name === "ValidationError") {
-      console.error(err);
-      return res
-        .status(Codes.Bad_Request)
-        .send({ message: "Указаны некорректные данные" });
+      return next(new BadRequestErr("Указаны некорректные данные"));
     }
-    console.error(err);
-    return res
-      .status(Codes.Internal_Server_Error)
-      .send({ message: "Ошибка на сервере. Добавить карточку не получилось." });
+    return next(err);
   }
 };
 
-const deleteCard = async (req, res) => {
+const deleteCard = async (req, res, next) => {
   const { cardId } = req.params;
+  const authUser = req.user._id;
   try {
-    const card = await Card.findByIdAndRemove(cardId);
+    const card = await Card.findById(cardId);
     if (card) {
-      return res.send({ message: `Карточка с id: ${cardId} удалена` });
+      if (owner === authUser) {
+        await Card.findByIdAndRemove(cardId);
+        return res.send({ message: `Карточка с id: ${cardId} удалена` });
+      }
+      throw new ForbiddenErr("Удаление чужих карточек запрещено");
     }
-    return res
-      .status(Codes.Not_Found)
-      .send({ message: `Карточка с id: ${cardId} не найдена` });
+    return next(new NotFoundErr(`Карточка с id: ${cardId} не найдена`));
   } catch (err) {
     if (err.name === "CastError") {
-      console.error(err);
-      return res
-        .status(Codes.Bad_Request)
-        .send({ message: "id карточки некорректный" });
+      return next(new BadRequestErr("id карточки некорректный"));
     }
-    console.error(err);
-    return res
-      .status(Codes.Internal_Server_Error)
-      .send({ message: "Произошла ошибка на сервере" });
+    return next(err);
   }
 };
 
-const likeCard = async (req, res) => {
+const likeCard = async (req, res, next) => {
   try {
     const { cardId } = req.params;
     const card = await Card.findByIdAndUpdate(
@@ -54,26 +46,18 @@ const likeCard = async (req, res) => {
       { new: true }
     );
     if (!card) {
-      return res
-        .status(Codes.Not_Found)
-        .send({ message: `Карточка с id: ${cardId} не найдена` });
+      throw new NotFoundErr(`Карточка с id: ${cardId} не найдена`);
     }
     return res.send({ likes: card.likes, message: "Лайк установлен" });
   } catch (err) {
     if (err.name === "CastError") {
-      console.error(err);
-      return res
-        .status(Codes.Bad_Request)
-        .send({ message: "id карточки некорректный" });
+      return next(new BadRequestErr("id карточки некорректный"));
     }
-    console.error(err);
-    return res
-      .status(Codes.Internal_Server_Error)
-      .send({ message: "Произошла ошибка на сервере" });
+    return next(err);
   }
 };
 
-const dislikeCard = async (req, res) => {
+const dislikeCard = async (req, res, next) => {
   try {
     const { cardId } = req.params;
     const card = await Card.findByIdAndUpdate(
@@ -82,34 +66,23 @@ const dislikeCard = async (req, res) => {
       { new: true }
     );
     if (!card) {
-      return res
-        .status(Codes.Not_Found)
-        .send({ message: `Карточка с id: ${cardId} не найдена` });
+      throw new NotFoundErr(`Карточка с id: ${cardId} не найдена`);
     }
     return res.send({ likes: card.likes, message: "Лайк удален" });
   } catch (err) {
     if (err.name === "CastError") {
-      console.error(err);
-      return res
-        .status(Codes.Bad_Request)
-        .send({ message: "id карточки некорректный" });
+      return next(new BadRequestErr("id карточки некорректный"));
     }
-    console.error(err);
-    return res
-      .status(Codes.Internal_Server_Error)
-      .send({ message: "Произошла ошибка на сервере" });
+    return next(err);
   }
 };
 
-const getCards = async (req, res) => {
+const getCards = async (req, res, next) => {
   try {
     const cards = await Card.find({});
     return res.json(cards);
   } catch (err) {
-    console.error(err);
-    return res
-      .status(Codes.Internal_Server_Error)
-      .send({ message: "Произошла ошибка на сервере" });
+    return next(err);
   }
 };
